aidoku::alloc::io

Module prelude

🔬This is a nightly-only experimental API. (alloc_io)

The I/O Prelude.

The purpose of this module is to alleviate imports of many common I/O traits by adding a glob import to the top of I/O heavy modules:

use alloc::io::prelude::*;

Traits

BufReadExperimental

A BufRead is a type of Reader which has an internal buffer, allowing it to perform extra ways of reading.

ReadExperimental

The Read trait allows for reading bytes from a source.

SeekExperimental

The Seek trait provides a cursor which can be moved within a stream of bytes.

WriteExperimental

A trait for objects which are byte-oriented sinks.