aidoku::alloc::io

Struct BorrowedBuf

pub struct BorrowedBuf<'data, T> { /* private fields */ }

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

A borrowed buffer of initially uninitialized elements, which is incrementally filled.

This type makes it safer to work with MaybeUninit buffers, such as to read into a buffer without having to initialize it first. It tracks the region of elements that have been filled and whether the unfilled region was initialized.

In summary, the contents of the buffer can be visualized as:

[                capacity                ]
[ filled | unfilled (may be initialized) ]

A BorrowedBuf is created around some existing elements (or capacity for elements) via a unique reference (&mut). The BorrowedBuf can be configured (e.g., using clear or set_init), but cannot be directly written. To write into the buffer, use unfilled to create a BorrowedCursor. The cursor has write-only access to the unfilled portion of the buffer (you can think of it as a write-only iterator).

The lifetime 'data is a bound on the lifetime of the underlying elements.

The type is most commonly used to manage bytes, but can manage any type of elements.

Implementations

impl<'data, T> BorrowedBuf<'data, T>

pub fn capacity(&self) -> usize

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns the total capacity of the buffer.

pub fn len(&self) -> usize

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns the length of the filled part of the buffer.

pub fn is_init(&self) -> bool

🔬This is a nightly-only experimental API. (borrowed_buf_init)

Returns true if the buffer is initialized.

impl<'data, T> BorrowedBuf<'data, T>
where T: Copy,

pub fn filled(&self) -> &[T]

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns a shared reference to the filled portion of the buffer.

pub fn filled_mut(&mut self) -> &mut [T]

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns a mutable reference to the filled portion of the buffer.

pub fn into_filled(self) -> &'data [T]

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns a shared reference to the filled portion of the buffer with its original lifetime.

pub fn into_filled_mut(self) -> &'data mut [T]

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns a mutable reference to the filled portion of the buffer with its original lifetime.

pub fn unfilled<'this>(&'this mut self) -> BorrowedCursor<'this, T>

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Returns a cursor over the unfilled part of the buffer.

pub fn clear(&mut self) -> &mut BorrowedBuf<'data, T>

🔬This is a nightly-only experimental API. (core_io_borrowed_buf)

Clears the buffer, resetting the filled region to empty.

The contents of the buffer are not modified.

pub unsafe fn set_init(&mut self) -> &mut BorrowedBuf<'data, T>

🔬This is a nightly-only experimental API. (borrowed_buf_init)

Asserts that the unfilled part of the buffer is initialized.

Safety

All the elements of the buffer must be initialized.