aidoku::alloc::io

Type Alias Result

pub type Result<T> = Result<T, Error>;

🔬This is a nightly-only experimental API. (alloc_io)

A specialized Result type for I/O operations.

This type is broadly used across std::io for any operation which may produce an error.

This type alias is generally used to avoid writing out io::Error directly and is otherwise a direct mapping to Result.

While usual Rust style is to import types directly, aliases of Result often are not, to make it easier to distinguish between them. Result is generally assumed to be core::result::Result, and so users of this alias will generally use io::Result instead of shadowing the prelude’s import of core::result::Result.

Examples

A convenience function that bubbles an io::Result to its caller:

use std::io;

fn get_string() -> io::Result<String> {
    let mut buffer = String::new();

    io::stdin().read_line(&mut buffer)?;

    Ok(buffer)
}

Aliased Type

pub enum Result<T> {
    Ok(T),
    Err(Error),
}

Variants

Ok(T)

🔬This is a nightly-only experimental API. (alloc_io)

Contains the success value

Err(Error)

🔬This is a nightly-only experimental API. (alloc_io)

Contains the error value