aidoku::canvas

Struct PathBuilder

pub struct PathBuilder { /* private fields */ }

A builder for paths.

Implementations

impl PathBuilder

pub fn new() -> Self

Creates a new path builder.

pub fn move_to(self, x: f32, y: f32) -> Self

Moves the current point to x, y.

pub fn line_to(self, x: f32, y: f32) -> Self

Adds a line segment from the current point to x, y.

pub fn quad_to(self, cx: f32, cy: f32, x: f32, y: f32) -> Self

Adds a quadratic bezier from the current point to x, y, using a control point of cx, cy.

pub fn rect(self, rect: &Rect) -> Self

Adds a rect to the path.

pub fn cubic_to( self, x: f32, y: f32, cx1: f32, cy1: f32, cx2: f32, cy2: f32, ) -> Self

Adds a cubic bezier from the current point to x, y, using control points cx1, cy1 and cx2, cy2.

pub fn arc( self, x: f32, y: f32, radius: f32, start_angle: f32, sweep_angle: f32, ) -> Self

Adds an arc approximated by quadratic beziers with center x, y and radius r starting at start_angle and sweeping by sweep_angle. For a positive sweep_angle the sweep is done clockwise, for a negative sweep_angle the sweep is done counterclockwise.

pub fn close(self) -> Self

Closes the current subpath.

pub fn build(self) -> Path

Builds the path.