aidoku::helpers::uri

Struct QueryParameters

pub struct QueryParameters { /* private fields */ }

Alternating, decoded query names and values.

Implementations

impl QueryParameters

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn push(&mut self, name: &str, value: Option<&str>)

Percent-encode the query paramter with encode_uri_component and add it to the query string along with a value.

pub fn push_key<K: AsRef<str>>(&mut self, name: K)

Percent-encode the query paramter with encode_uri_component and add it to the query string.

pub fn push_encoded(&mut self, name: &str, value: Option<&str>)

Add a pre-encoded query parameter to the query string.

pub fn set(&mut self, name: &str, value: Option<&str>)

Percent-encode the query parameter with encode_uri_component and replace any existing values.

pub fn set_encoded(&mut self, name: &str, value: Option<&str>)

Replace any existing values with the given pair, without encoding.

pub fn remove_all<T: AsRef<str>>(&mut self, name: T)

Remove all query parameters matching given name.

pub fn remove_all_encoded<T: AsRef<str>>(&mut self, name: T)

Remove all query parameters matching given pre-encoded name.

pub fn from_data<T: Serialize>(value: &T) -> Result<Self, SerializeError>

Serialize the given data as a QueryParameters struct.