aidoku::imports::canvas

Struct ImageRef

pub struct ImageRef {
    pub rid: i32,
    pub externally_managed: bool,
}

A reference to an image.

Fields
rid: i32

The reference id of the stored image.

This property is exposed for the functions that the register_source macro generates and should not be used directly.

externally_managed: bool

Whether the image is externally managed.

This property is exposed for the functions that the register_source macro generates and should not be used directly.

Implementations

impl ImageRef

pub fn new(data: &[u8]) -> Self

pub fn data(&self) -> Vec<u8>

pub fn width(&self) -> f32

pub fn height(&self) -> f32