aidoku::imports::html

Struct Html

pub struct Html;

Namespace for HTML-related functions.

Implementations

impl Html

pub fn parse<T: AsRef<[u8]>>(html: T) -> Result<Document, HtmlError>

Parse HTML into a Document.

As there is no base URL specified, absolute URL resolution requires the HTML to have a <base href> tag.

pub fn parse_with_url<T: AsRef<[u8]>, B: AsRef<str>>( html: T, base_url: B, ) -> Result<Document, HtmlError>

Parse HTML into a Document, with a base URL.

The given base_url will be used for any URLs that occurs before a <base href> tag is defined.

pub fn parse_fragment<T: AsRef<[u8]>>(html: T) -> Result<Document, HtmlError>

Parse a HTML fragment, assuming that it forms the body of the HTML.

Similar to Html::parse, relative URLs will not be resolved unless there is a <base href> tag.

pub fn parse_fragment_with_url<T: AsRef<[u8]>, B: AsRef<str>>( html: T, base_url: B, ) -> Result<Document, HtmlError>

Parse a HTML fragment, assuming that it forms the body of the HTML, with a base URL.

Similar to Html::parse_with_url, URL resolution occurs for any that appears before a <base href> tag.

pub fn escape<T: AsRef<str>>(text: T) -> String

Escape any HTML-reserved characters to HTML entities.

Examples

use aidoku::imports::html::Html;
assert_eq!(
    Html::escape("Hello &<> Å å π 新 there ¾ © »"),
    "Hello &amp;&lt;&gt; Å å π 新 there ¾ © »",
);

pub fn unescape<T: AsRef<str>>(text: T) -> Option<String>

Unescape any HTML entities to their original characters.

Examples

use aidoku::imports::html::Html;
assert_eq!(
    Html::unescape("Hello &amp;&lt;&gt; Å å π 新 there ¾ © »"),
    Some("Hello &<> Å å π 新 there ¾ © »".into()),
);