aidoku::imports::html

Struct Node

pub struct Node { /* private fields */ }

A single HTML node.

Implementations

impl Node

pub fn kind(&self) -> Kind

Get the kind of the node.

pub fn parent(&self) -> Option<Node>

Get the node’s parent node, returning None if there isn’t one.

pub fn child_nodes(&self) -> Vec<Node>

Get the node’s child nodes.

pub fn siblings(&self) -> Vec<Node>

Get the sibling nodes of the node.

pub fn next(&self) -> Option<Node>

Get the next sibling of the node, returning None if there isn’t one.

pub fn prev(&self) -> Option<Node>

Get the previous sibling of the node, returning None if there isn’t one.

pub fn outer_html(&self) -> Option<String>

Get the node’s outer HTML.

pub fn has_attr<T: AsRef<str>>(&self, attr_name: T) -> bool

Test if this node has an attribute. Case insensitive.

pub fn set_attr<K: AsRef<str>, V: AsRef<str>>( &mut self, key: K, value: V, ) -> Result<(), HtmlError>

Set an attribute value on this node.

If this node already has an attribute with the key, its value is updated; otherwise, a new attribute is added.

pub fn remove_attr<T: AsRef<str>>(&mut self, attr: T) -> Result<(), HtmlError>

Remove an attribute from this node.

pub fn text(&self) -> Option<String>

Get the text of this node, if it is a text node.

pub fn data(&self) -> Option<String>

Get the data of this node, if it is a data node or comment.