aidoku::imports::net

Struct Request

pub struct Request {
    pub rid: i32,
    pub should_close: bool,
    /* private fields */
}

An HTTP request.

Fields
rid: i32

The reference id for the request.

This property is exposed for the functions that the register_source macro generates and should not be used directly.

should_close: bool

Whether the request should be closed after being dropped.

This property is exposed for the functions that the register_source macro generates and should not be used directly.

Implementations

impl Request

pub fn new<T: AsRef<str>>( url: T, http_method: HttpMethod, ) -> Result<Self, RequestError>

Create a new request with a URL and HTTP method.

Returns an error if the provided URL is invalid.

Examples

use aidoku::imports::net::{HttpMethod, Request};
Request::new("https://example.com", HttpMethod::Get).unwrap();

pub fn get<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new GET request with the given URL.

pub fn post<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new POST request with the given URL.

pub fn put<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new PUT request with the given URL.

pub fn head<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new HEAD request with the given URL.

pub fn delete<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new DELETE request with the given URL.

pub fn patch<T: AsRef<str>>(url: T) -> Result<Self, RequestError>

Create a new PATCH request with the given URL.

pub fn send_all<I>(requests: I) -> Vec<Result<Response, RequestError>>
where I: IntoIterator<Item = Request>,

Send multiple requests in parallel, and wait for all of them to finish.

pub fn header<T: AsRef<str>>(self, key: T, val: T) -> Self

Set an HTTP header in a builder.

pub fn set_header<T: AsRef<str>>(&mut self, key: T, val: T)

Set an HTTP header.

pub fn body<T: AsRef<[u8]>>(self, data: T) -> Self

Set the HTTP body data in a builder.

pub fn timeout(self, value: f64) -> Self

Set the request timeout interval in a builder.

The request timeout interval controls how long (in seconds) a task should wait for additional data to arrive before giving up.

pub fn set_body<T: AsRef<[u8]>>(&mut self, data: T)

Set the HTTP body data.

pub fn set_timeout(&mut self, value: f64)

Set the request timeout interval.

The request timeout interval controls how long (in seconds) a task should wait for additional data to arrive before giving up.

pub fn set_url<T: AsRef<str>>(&mut self, url: T) -> Result<(), RequestError>

Set the URL for the request.

pub fn url(&self) -> Option<&String>

Get the URL of the request.

pub fn send(self) -> Result<Response, RequestError>

Send the request.

pub fn data(self) -> Result<Vec<u8>, RequestError>

Get the raw data from the response, closing the request.

pub fn image(self) -> Result<ImageRef, RequestError>

Gets the response data as an image.

pub fn string(self) -> Result<String, AidokuError>

Gets the response data as a string.

pub fn html(self) -> Result<Document, RequestError>

Get the response data as an HTML Document.