aidoku::imports::std

Function send_partial_result

pub fn send_partial_result<T: Serialize>(value: &T)

Sends a partial result to the source runner.

This function is used to send partial home layours and manga results back to the runner for faster loading.

Only HomePartialResult and Manga structs should be passed as arguments.