Struct aidoku::std::ValueRef

pub struct ValueRef(pub i32, pub bool);

A type which can represent value of any kind. It is used when exchanging information with the Aidoku app.

Tuple Fields
0: i321: bool
Implementations

impl ValueRef

pub fn new(rid: i32) -> ValueRef

Create a new ValueRef which points to a descriptor.

pub fn kind(&self) -> Kind

Get the type of the ValueRef.

pub fn is_none(&self) -> bool

Check if the ValueRef is null.

pub fn is_some(&self) -> bool

Check if the ValueRef is not null.

pub fn as_string(self) -> Result<StringRef, AidokuError>

Cast the ValueRef to a StringRef.

pub fn as_object(self) -> Result<ObjectRef, AidokuError>

Cast the ValueRef to an ObjectRef.

pub fn as_array(self) -> Result<ArrayRef, AidokuError>

Cast the ValueRef to an ArrayRef.

pub fn as_int(&self) -> Result<i64, AidokuError>

Cast the ValueRef to an i64.

pub fn as_float(&self) -> Result<f64, AidokuError>

Cast the ValueRef to an f64.

pub fn as_bool(&self) -> Result<bool, AidokuError>

Cast the ValueRef to a boolean.

pub fn as_date<T>( &self, format: T, locale: Option<T>, timezone: Option<T> ) -> Result<f64, AidokuError>where T: Default + AsRef<str>,

Converts a textual representaion of a date to a Unix timestamp.

Arguments

format: The date format, as compatible with NSDateFormatter.

locale: The locale identifier for this date string. Also available on NSDateFormatter.

timezone: The time zone for this date, as compatible with TimeZone. They can be a zoneinfo timezone, or an abbreviation.

pub fn as_node(&self) -> Result<Node, AidokuError>

Cast the ValueRef to a Node.