aidoku

Struct MultiSelectFilter

pub struct MultiSelectFilter {
    pub id: Cow<'static, str>,
    pub title: Option<Cow<'static, str>>,
    pub hide_from_header: Option<bool>,
    pub is_genre: bool,
    pub can_exclude: bool,
    pub uses_tag_style: bool,
    pub options: Vec<Cow<'static, str>>,
    pub ids: Option<Vec<Cow<'static, str>>>,
    pub default_included: Option<Vec<Cow<'static, str>>>,
    pub default_excluded: Option<Vec<Cow<'static, str>>>,
}

A list of values that allows multiple selections.

This filter can be converted into a generic Filter using the Into trait.

Fields
id: Cow<'static, str>

The identifier for this filter.

title: Option<Cow<'static, str>>

The display title for this filter.

hide_from_header: Option<bool>

Whether this filter should be hidden from the filters list header.

is_genre: bool

Indicates if the filter is for genres.

can_exclude: bool

Whether options can be excluded as well as included.

uses_tag_style: bool

Whether to display the options as tags.

options: Vec<Cow<'static, str>>

The list of options to display.

ids: Option<Vec<Cow<'static, str>>>

Optional IDs for each option. If not provided, the options are used.

default_included: Option<Vec<Cow<'static, str>>>default_excluded: Option<Vec<Cow<'static, str>>>