aidoku

Trait Source

pub trait Source {
    // Required methods
    fn new() -> Self;
    fn get_search_manga_list(
        &self,
        query: Option<String>,
        page: i32,
        filters: Vec<FilterValue>,
    ) -> Result<MangaPageResult>;
    fn get_manga_update(
        &self,
        manga: Manga,
        needs_details: bool,
        needs_chapters: bool,
    ) -> Result<Manga>;
    fn get_page_list(&self, manga: Manga, chapter: Chapter) -> Result<Vec<Page>>;
}

The required functions an Aidoku source must implement.

Required Methods

fn new() -> Self

Called to initialize a source.

If a source requires any setup before other functions are called, it should happen here.

fn get_search_manga_list( &self, query: Option<String>, page: i32, filters: Vec<FilterValue>, ) -> Result<MangaPageResult>

Returns the manga for a search query with filters.

fn get_manga_update( &self, manga: Manga, needs_details: bool, needs_chapters: bool, ) -> Result<Manga>

Updates a given manga with new details and chapters, as requested.

fn get_page_list(&self, manga: Manga, chapter: Chapter) -> Result<Vec<Page>>

Returns the pages for a given manga chapter.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.