Crate aidoku

Aidoku Rust Source API

NOTE: This library currently requires Rust nightly, which can be installed with rustup default nightly.

Setup

To make a new source:

cargo new <source_name>

Then in the Cargo.toml add:

[lib]
crate-type = ["cdylib"]

[profile.dev]
panic = "abort"

[profile.release]
panic = "abort"
opt-level = "s"
strip = true
lto = true

[dependencies]
aidoku = "0.1.0"

Next, make a folder called .cargo and a file called “config”. In it, put:

[build]
target = "wasm32-unknown-unknown"

Now, in your src/main.rs file, simply add:

#![no_std]
use aidoku::{
    error::Result,
    prelude::*,
    std::{String, Vec},
    Chapter, Filter, Listing, Manga, MangaPageResult, Page,
};

#[get_manga_list]
fn get_manga_list(_: Vec<Filter>, _: i32) -> Result<MangaPageResult> {
    todo!()
}

#[get_manga_listing]
fn get_manga_listing(_: Listing, _: i32) -> Result<MangaPageResult> {
    todo!()
}

#[get_manga_details]
fn get_manga_details(_: String) -> Result<Manga> {
    todo!()
}

#[get_chapter_list]
fn get_chapter_list(_: String) -> Result<Vec<Chapter>> {
    todo!()
}

#[get_page_list]
fn get_page_list(_: String) -> Result<Vec<Page>> {
    todo!()
}

And now you’re ready to start making your source!

Modules

error

Error module for Aidoku operations.

prelude

The Aidoku prelude, which includes format!, println!, as well as procedural macros which are required for interop with the app.

std

The Aidoku standard module, which includes all functions exported from Aidoku as well as a few common types.

Structs

Chapter

Struct containing metadata about a chapter. Different chapters are differentiated by their ID. Thus, changing the ID will result in a different chapter, even if the chapters have the same volume/chapter number.

DeepLink

Struct representing a deep link. This deep link is used to open a manga from a webpage URL using the aidoku:// URL scheme.

Filter

Struct representing a search filter.

Listing

Struct containing information about a listing.

Manga

The Manga struct contains information about a manga. Different mangas are differentiated by their ID, and so changing the ID will result in a different manga. Thus, developers should decide on the ID format before publishing their source.

MangaPageResult

A struct representing a “page” of mangas. There is no limit on how many mangas can a page have, that is up to the source to decide.

Page

Struct representing a manga page.

Enums

FilterType

MangaContentRating

An enumeration representing the manga’s content rating.

MangaStatus

An enum representing the various statuses a manga can have.

MangaViewer

An enumeration representing different manga viewers, used to indicate the preferred reading method for this manga.

Functions

alloc_error_handle

panic_handle