Module aidoku::std

The Aidoku standard module, which includes all functions exported from Aidoku as well as a few common types.

Modules

defaults

Interface for interacting with user preferences.

error

html

A module for working with HTML. It provides a somewhat convenient API for extracting data, using HTML5 DOM methods and CSS selectors.

json

Converts JSON to a ValueRef.

net

Create and send HTTP requests.

Structs

ArrayRef

A type which represents an array.

ObjectRef

A type that represents a string-keyed and value object.

String

A UTF-8–encoded, growable string.

StringRef

A type which represents a string.

ValueRef

A type which can represent value of any kind. It is used when exchanging information with the Aidoku app.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

Kind

An enumeration of different ValueRef types.

Traits

Write

A trait for writing or formatting into Unicode-accepting buffers or streams.

Functions

copy⚠

Copies a ValueRef and returns a descriptor pointing to the new ValueRef.

current_date

Gets the current time as a Unix timestamp.

destroy⚠

Destroys a ValueRef.

format

print

Prints a message to the Aidoku logs.

value_kind⚠

Returns the kind of the ValueRef.

Type Definitions

Rid

A standard descriptor, used for data exchange between the app and the source.